Expose "colors" dictionary
My ultimate plan with this in Blitz Search is to provide "easy" theme selection.  But I do think that AvaloniaEdit.TextMate should use the editor background to fix what I see as a bug in the AvaloniaEdit demonstration, where some themes don't look like they were meant for a dark background.

Having things like selection highlights and all that come from a place that's already authored will ease up a lot of UI work for me ( color pickers and what-not ).